Updated wordpress__rich-text to 6.0
