+++ types/wordpress__rich-text/component/use-anchor.d.ts
@@ -1,0 +1,8 @@
+import { Format, Value } from "../create";
+
+export interface VirtualAnchorElement {
+    getBoundingClientRect(): DOMRect;
+    ownerDocument: Document;
+}
+
+export function useAnchor(param: {editableContentElement: HTMLElement|null, value: Value, settings?: Format}): Element|VirtualAnchorElement|undefined|null;
